Rewritten ckeditor5-basic-styles to TypeScript.
