+++ .gitignore
@@ -25,8 +25,8 @@ mrgit.json
 .yalc
 yalc.lock
 
-
 # Ignore compiled TypeScript files.
+packages/ckeditor5-basic-styles/src/**/*.js
 packages/ckeditor5-clipboard/src/**/*.js
 packages/ckeditor5-core/src/**/*.js
 packages/ckeditor5-engine/src/**/*.js
