+++ packages/ckeditor5-basic-styles/src/index.ts
