+++ packages/ckeditor5-basic-styles/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "ckeditor5/tsconfig.json",
+  "include": [
+    "src",
+    "../../typings"
+  ]
+}
